+++ test/weak_convergence/srk_weak_final.jl
@@ -311,139 +311,6 @@ m = log(errors[end]/errors[1])/log(dts[end]/dts[1])
 println("RDI4WM:", m)
 
 """
- Test non-commutative noise SDEs (iip)
-"""
-
-@info "Non-commutative noise"
-
-u₀ = [1.0,1.0]
-function f2!(du,u,p,t)
-  du[1] = -273//512*u[1]
-  du[2] = -1//160*u[1]-(-785//512+sqrt(2)/8)*u[2]
-end
-function g2!(du,u,p,t)
-  du[1,1] = 1//4*u[1]
-  du[1,2] = 1//16*u[1]
-  du[2,1] = (1-2*sqrt(2))/4*u[1]
-  du[2,2] = 1//10*u[1]+1//16*u[2]
-end
-dts = 1 .//2 .^(3:-1:0)
-tspan = (0.0,3.0)
-
-h2(z) = z^2 # but apply it only to u[1]
-
-prob = SDEProblem(f2!,g2!,u₀,tspan,noise_rate_prototype=zeros(2,2))
-ensemble_prob = EnsembleProblem(prob;
-        output_func = (sol,i) -> (h2(sol[end][1]),false),
-        prob_func = prob_func
-        )
-
-numtraj = Int(1e5)
-seed = 100
-Random.seed!(seed)
-seeds = rand(UInt, numtraj)
-
-
-_solutions = @time generate_weak_solutions(ensemble_prob, DRI1(), dts, numtraj, ensemblealg=EnsembleThreads())
-
-errors = [LinearAlgebra.norm(Statistics.mean(sol.u)-exp(-3.0)) for sol in _solutions]
-m = log(errors[end]/errors[1])/log(dts[end]/dts[1])
-@test -(m-2) < 0.3
-
-println("DRI1:", m)
-
-numtraj = Int(2e5)
-seed = 100
-Random.seed!(seed)
-seeds = rand(UInt, numtraj)
-_solutions = @time generate_weak_solutions(ensemble_prob, RI1(), dts, numtraj, ensemblealg=EnsembleThreads())
-
-errors = [LinearAlgebra.norm(Statistics.mean(sol.u)-exp(-3.0)) for sol in _solutions]
-m = log(errors[end]/errors[1])/log(dts[end]/dts[1])
-@test -(m-2) < 0.3
-
-println("RI1:", m)
-
-numtraj = Int(2e5)
-seed = 100
-Random.seed!(seed)
-seeds = rand(UInt, numtraj)
-_solutions = @time generate_weak_solutions(ensemble_prob, RI3(), dts, numtraj, ensemblealg=EnsembleThreads())
-
-errors = [LinearAlgebra.norm(Statistics.mean(sol.u)-exp(-3.0)) for sol in _solutions]
-m = log(errors[end]/errors[1])/log(dts[end]/dts[1])
-@test -(m-2) < 0.3
-
-println("RI3:", m)
-
-numtraj = Int(2e5)
-seed = 100
-Random.seed!(seed)
-seeds = rand(UInt, numtraj)
-_solutions = @time generate_weak_solutions(ensemble_prob, RI5(), dts, numtraj, ensemblealg=EnsembleThreads())
-
-errors = [LinearAlgebra.norm(Statistics.mean(sol.u)-exp(-3.0)) for sol in _solutions]
-m = log(errors[end]/errors[1])/log(dts[end]/dts[1])
-@test -(m-2) < 0.3
-
-println("RI5:", m)
-
-numtraj = Int(2e5)
-seed = 100
-Random.seed!(seed)
-seeds = rand(UInt, numtraj)
-_solutions = @time generate_weak_solutions(ensemble_prob, RI6(), dts, numtraj, ensemblealg=EnsembleThreads())
-
-errors = [LinearAlgebra.norm(Statistics.mean(sol.u)-exp(-3.0)) for sol in _solutions]
-m = log(errors[end]/errors[1])/log(dts[end]/dts[1])
-@test -(m-2) < 0.3
-
-println("RI6:", m)
-
-
-numtraj = Int(1e3)
-seed = 10
-Random.seed!(seed)
-seeds = rand(UInt, numtraj)
-
-_solutions = @time generate_weak_solutions(ensemble_prob, RDI1WM(), dts, numtraj, ensemblealg=EnsembleThreads())
-
-errors = [LinearAlgebra.norm(Statistics.mean(sol.u)-exp(-3.0)) for sol in _solutions]
-m = log(errors[end]/errors[1])/log(dts[end]/dts[1])
-@test -(m-1) < 0.3
-
-println("RDI1WM:", m)
-
-numtraj = Int(1e5)
-seed = 10
-Random.seed!(seed)
-seeds = rand(UInt, numtraj)
-
-_solutions = @time generate_weak_solutions(ensemble_prob, RDI2WM(), dts, numtraj, ensemblealg=EnsembleThreads())
-
-errors = [LinearAlgebra.norm(Statistics.mean(sol.u)-exp(-3.0)) for sol in _solutions]
-m = log(errors[end]/errors[1])/log(dts[end]/dts[1])
-@test -(m-2) < 0.3
-
-println("RDI2WM:", m)
-
-_solutions = @time generate_weak_solutions(ensemble_prob, RDI3WM(), dts, numtraj, ensemblealg=EnsembleThreads())
-
-errors = [LinearAlgebra.norm(Statistics.mean(sol.u)-exp(-3.0)) for sol in _solutions]
-m = log(errors[end]/errors[1])/log(dts[end]/dts[1])
-@test -(m-2) < 0.3
-
-println("RDI3WM:", m)
-
-_solutions = @time generate_weak_solutions(ensemble_prob, RDI4WM(), dts, numtraj, ensemblealg=EnsembleThreads())
-
-errors = [LinearAlgebra.norm(Statistics.mean(sol.u)-exp(-3.0)) for sol in _solutions]
-m = log(errors[end]/errors[1])/log(dts[end]/dts[1])
-@test -(m-2) < 0.3
-
-println("RDI4WM:", m)
-
-"""
  Test Diagonal noise SDEs (iip), SIAM Journal on Numerical Analysis, 47 (2009), pp. 1713–1738
 """
